(PC-25795) refactor(offer): use hook in offer feature instead of offerv2 to avoid dead code
