+++ src/features/offer/components/OfferBody/OfferBody.tsx
@@ -12,8 +12,8 @@ import { OfferPartialDescription } from 'features/offer/components/OfferPartialD
 import { OfferPlace } from 'features/offer/components/OfferPlace/OfferPlace'
 import { HitOfferWithArtistAndEan } from 'features/offer/components/OfferPlaylist/api/fetchOffersByArtist'
 import { OfferPlaylistList } from 'features/offer/components/OfferPlaylistList/OfferPlaylistList'
+import { useOfferBodyData } from 'features/offer/helpers/useOfferBodyData/useOfferBodyData'
 import { useTrackOfferSeenDuration } from 'features/offer/helpers/useTrackOfferSeenDuration'
-import { useOfferBodyData } from 'features/offerv2/helpers/useOfferBodyData/useOfferBodyData'
 import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
 import { analytics } from 'libs/analytics'
 import { useLocation } from 'libs/location'
