+++ src/features/offer/components/OfferContent/OfferContent.tsx
@@ -14,9 +14,9 @@ import { OfferBody } from 'features/offer/components/OfferBody/OfferBody'
 import { OfferCTAButton } from 'features/offer/components/OfferCTAButton/OfferCTAButton'
 import { OfferHeader } from 'features/offer/components/OfferHeader/OfferHeader'
 import { OfferWebHead } from 'features/offer/components/OfferWebHead'
-import { useOfferAnalytics } from 'features/offerv2/helpers/useOfferAnalytics/useOfferAnalytics'
-import { useOfferBatchTraking } from 'features/offerv2/helpers/useOfferBatchTracking/useOfferBatchTracking'
-import { useOfferPlaylist } from 'features/offerv2/helpers/useOfferPlaylist/useOfferPlaylist'
+import { useOfferAnalytics } from 'features/offer/helpers/useOfferAnalytics/useOfferAnalytics'
+import { useOfferBatchTraking } from 'features/offer/helpers/useOfferBatchTracking/useOfferBatchTracking'
+import { useOfferPlaylist } from 'features/offer/helpers/useOfferPlaylist/useOfferPlaylist'
 import { isCloseToBottom } from 'libs/analytics'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
